COVID Casr U.S. Trend analysis
COVID-19 is a novel coronavirus, which caused pandemic in the beginning of 2020. The dataset to be analyzed, is the cases reported in the United States, is provided by Johns Hopkins University. There are two datasets (covid_us_county.csv and us_county.csv) that can be downloaded from Kaggle, containing information about total COVID-19 cases and death by county and gender.
The analysis was done as part of Data Analytics course at Careerfoundry.